+++ dd-front-end/src/App.jsx
@@ -1,9 +1,9 @@
 import React from "react";
-import RegistrationPage from "./components/RegistrationPage";
+import RegistrationPage from "./components/Registration/RegistrationPage";
 import RegPage from "./components/RegPage";
 import OtpAuth from "./components/OtpAuth";
 import { Route, Routes } from "react-router-dom";
-import LoginPage from "./components/LoginPage";
+import LoginPage from "./components/Login/LoginPage";
 
 export default function App() {
   return (
