--- dd-front-end/src/components/InputField.jsx
@@ -1,15 +1,0 @@
-
-const InputField = (props) => {
-    return (
-        <input
-            className="input--field"
-            id={props.id}
-            type={props.type}
-            value={props.value}
-            placeholder={props.placeholder}
-        // onChange={props.onChangeHandler}
-        />
-    )
-}
-
-export default InputField
