--- dd-front-end/src/components/RegPage.jsx
@@ -1,25 +1,0 @@
-import '../styles/regPage.css'
-
-const RegPage = () => {
-    return (
-        <div id='main--container'>
-            <div className='blocks' id='left--block'>
-                <div id='title'>Register to use DirectDealz!</div>
-                <div id='form--div'>
-                    <form>
-                    </form>
-                </div>
-                <div id='sac--button--div'>
-                    <button>Save & Continue</button>
-                </div>
-            </div>
-            <div className='blocks' id='right--block'>
-                <div>Figurine</div>
-                <div>Form</div>
-                <div>Button</div>
-            </div>
-        </div>
-    )
-}
-
-export default RegPage
