--- dd-front-end/src/styles/regPage.css
@@ -1,53 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
-
-#main--container {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    width: 50%;
-    height: 60%;
-    background-color: antiquewhite;
-}
-
-.blocks {
-    position: absolute;
-    top: 5%;
-    width: 45%;
-    height: 90%;
-    background-color: pink;
-}
-
-#left--block {
-    left: 3.75%;
-}
-
-#right--block {
-    right: 3.75%;
-}
-
-#left--block>div {
-    width: 100%;
-}
-
-#title {
-    position: absolute;
-    height: 20%;
-    background-color: aliceblue;
-}
-
-#form--div {
-    position: absolute;
-    height: 65%;
-}
-
-#sac--button--div {
-    position: absolute;
-    bottom: 0;
-    height: 15%;
-    background-color: bisque;
-}
